test(#2760): move dispatch-idle REDs to LOC-exempt sibling (anti-monolith ceiling)
The full all-target run flagged src_file_size_invariant: the 2 inline dispatch-idle REDs
grew grandfathered dispatch_idle/mod.rs from 3962→4022 LOC. Moved them to
tests/dispatch_idle_reds_2760_tests.rs, declared #[path] inside mod tests (the same
LOC-exempt split pattern as dispatch_idle_quota_78445_tests.rs) — as a submodule of
mod tests, use super::* inherits the inline helpers + production items. mod.rs back to
3955 LOC (< 3962). Invariant green; REDs 2/2 green; dispatch_idle 66/66.


Agend-Agent: archfix-opus-2
Agend-Task: t-20260712233217991202-15764-1
Agend-Branch: fix/project-board-strict-routing
Agend-Issued-At: 2026-07-13T11:22:01.415010+00:00
